Add tests for read_trial, separate run and trial tests
